Add a module for creating tasks; edit index.js to add event listeners.
