+++ src/adddelupd.js
@@ -1,0 +1,9 @@
+function createTaskElement(taskName, tasksLocal) {
+  const index = tasksLocal.length + 1;
+  const complete = false;
+  const taskString = { index, name: taskName, completed: complete };
+  tasksLocal.push(taskString);
+  localStorage.setItem('tasks', JSON.stringify(tasksLocal));
+}
+
+export default createTaskElement;
