+++ src/index.js
@@ -1,15 +1,13 @@
 import './style.css';
-/* global loadTasksFromLocalStorage */
 
+import createTaskElement from './adddelupd.js';
+
+/* global tasksLocal */
 let tasksLocal = [];
 
 window.loadTasksToLocalStorage = () => {
   const text = JSON.stringify(tasksLocal);
   localStorage.setItem('tasks', text);
-};
-
-window.loadTasksFromLocalStorage = () => {
-  tasksLocal = localStorage.getItem('tasks') ? JSON.parse(localStorage.getItem('tasks')) : [];
 };
 
 const displayTaskElement = (task) => {
@@ -36,6 +34,7 @@ const displayTaskElement = (task) => {
 
   const deleteIcon = document.createElement('span');
   deleteIcon.classList.add('delete-icon');
+  // deleteIcon.classList.add('hide');
   deleteIcon.innerHTML = '&#128465;';
 
   taskItem.appendChild(taskIndex);
@@ -44,22 +43,35 @@ const displayTaskElement = (task) => {
   taskItem.appendChild(moreIcon);
   taskItem.appendChild(deleteIcon);
 
+  activateMoreListeners();
+  activateDeleteListeners();
+  // activateInputListeners();
+
   return taskItem;
 };
 
-function createTaskElement(taskName) {
-  const index = tasksLocal.length + 1;
-  const complete = false;
-  const taskString = { index, name: taskName, completed: complete };
-  tasksLocal.push(taskString);
-  localStorage.setItem('tasks', JSON.stringify(tasksLocal));
-}
+function activateMoreListeners() {
+  let moreBtn = document.querySelectorAll('.more-icon');
+  moreBtn.forEach((mb, taskIndex) => {
+    mb.addEventListener('click', () => {
+      deleteIcon(taskIndex).classList.remove('hide');
+    });
+  });
+};
+
+// function createTaskElement(taskName) {
+//   const index = tasksLocal.length + 1;
+//   const complete = false;
+//   const taskString = { index, name: taskName, completed: complete };
+//   tasksLocal.push(taskString);
+//   localStorage.setItem('tasks', JSON.stringify(tasksLocal));
+// }
 
 document.getElementById('add-task-btn').addEventListener('click', () => {
   const taskInput = document.getElementById('task-input');
   const taskName = taskInput.value.trim();
   if (taskName !== '') {
-    createTaskElement(taskName);
+    createTaskElement(taskName, tasksLocal);
     document.location.reload();
     taskInput.value = '';
   }
@@ -73,6 +85,10 @@ const displayTasks = () => {
       taskList.appendChild(taskElement);
     });
   }
+};
+
+const loadTasksFromLocalStorage = () => {
+  tasksLocal = localStorage.getItem('tasks') ? JSON.parse(localStorage.getItem('tasks')) : [];
 };
 
 window.onload = () => {
